Fix for pages being deleted upon creation
The onClick method being assigned to the delete button
'this.deletePage(i)' was being called immediately upon rendering.
Changed it so that it was only called on click, as intended.

Also changed instance of 'Route' to 'Router' in the createPage method,
as it was giving me an error and not updating the page right away.